Organise files into packages
