+++ src/main/java/Main.java
@@ -1,0 +1,10 @@
+import java.io.IOException;
+
+import pelopsii.PelopsII;
+
+
+public class Main {
+    public static void main(String[] args) throws IOException {
+        new PelopsII("./src/main/java/data").run();
+    }
+}
