+++ src/main/java/pelopsii/exception/PelopsIIException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package pelopsii.exception;
 public class PelopsIIException extends Exception {
     public PelopsIIException(String message) {
         super(message);
