+++ src/main/java/pelopsii/storage/Storage.java
@@ -1,4 +1,4 @@
-package storage;
+package pelopsii.storage;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
@@ -6,9 +6,8 @@ import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
-import Exceptions.PelopsIIException;
-// import Tasks.Task;
-// import Tasks.TaskList;
+
+import pelopsii.exception.PelopsIIException;
 
 public class Storage {
     private final String filePath;
