+++ src/main/java/pelopsii/task/Deadline.java
@@ -1,4 +1,4 @@
-package Tasks;
+package pelopsii.task;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
