+++ src/main/java/pelopsii/task/Event.java
@@ -1,4 +1,4 @@
-package Tasks;
+package pelopsii.task;
 public class Event extends Task {
 
     protected String from;
