+++ src/main/java/pelopsii/ui/Ui.java
@@ -1,4 +1,4 @@
-package ui;
+package pelopsii.ui;
 
 import java.io.BufferedReader;
 import java.io.IOException;
